+++ CHANGELOG.md
@@ -38,6 +38,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - Disable and spinner to Challenge Task delete button
 - Disable and spinner to sign in submit buttons
 - Mission control button to update user triggers
+- Redeem all button for household vendor codes
+- Participant age to Group Vendor Code Report
 
 ### Changed
 - Script docker-build.bash to not build release images unless the branch starts with 'release/'